requirements.txt
playwright
colorama
customtkinter
Pillow
